Split jobs-pane.tsx into focused tab and overview components (#649)
